CountedHashSet - Cleanup and Docs
